deepshield-backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
motor==3.3.1

# AI and ML dependencies
torch==2.1.1
transformers==4.35.2
facenet-pytorch==2.5.3
deepface==0.0.79
hatesonar==0.0.10
nudenet==2.0.9
datasets==2.15.0

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
